EcommerceMCPServer/pyproject.toml
[project]
name = "ecommercemcpserver"
version = "0.1.0"
description = "An MCP server for an e-commerce store that exposes cart operations as MCP tools, and a client that leverages the LLM to interact with these tools."
readme = "README.md"
requires-python = "==3.13.7"
dependencies = [
    "asyncio>=4.0.0",
    "click>=8.2.1",
    "google-adk>=1.13.0",
    "mcp>=1.13.1",
    "python-dotenv>=1.1.1",
]
